requirements.txt
# =============================================================================
# Medical Claims Paid Amount Prediction - Requirements
# =============================================================================

# Core Data Science
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<3.0.0
scipy>=1.9.0,<2.0.0

# Machine Learning
scikit-learn>=1.2.0,<2.0.0
xgboost>=1.7.0,<3.0.0
lightgbm>=3.3.0,<5.0.0

# Visualization
matplotlib>=3.6.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Data Storage & Formats
pyarrow>=10.0.0,<18.0.0
fastparquet>=2023.1.0

# Configuration
PyYAML>=6.0,<7.0.0

# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# Utilities
psutil>=5.9.0,<6.0.0
tqdm>=4.64.0,<5.0.0
joblib>=1.2.0,<2.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.20.0
notebook>=6.5.0

# Development
pytest>=7.2.0,<9.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: Advanced ML (uncomment if needed)
# pyearth>=0.1.0  # MARS models
# shap>=0.41.0    # Model explanations
# optuna>=3.0.0   # Hyperparameter optimization
